gav-autonomo/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic>=2.0
pydantic-settings>=2.0
PyYAML==6.0.1
jsonschema==4.22.0
httpx[http2]